chore: remove UnityParcelScene from dcl.ts and initializeEngine
