+++ kernel/packages/entryPoints/editor.ts
@@ -12,7 +12,7 @@ import { SceneJsonData, ILand, HUDElementID } from 'shared/types'
 import { normalizeContentMappings } from 'shared/selectors'
 import { SceneWorker } from 'shared/world/SceneWorker'
 import { initializeUnity } from 'unity-interface/initializer'
-import { UnityParcelScene } from 'unity-interface/dcl'
+import { UnityParcelScene } from 'unity-interface/UnityParcelScene'
 import { loadBuilderScene, updateBuilderScene, unloadCurrentBuilderScene } from 'unity-interface/builder'
 import defaultLogger from 'shared/logger'
 import { uuid } from 'decentraland-ecs/src/ecs/helpers'
