+++ kernel/packages/unity-interface/initializer.ts
@@ -3,7 +3,7 @@ import { DEBUG_MESSAGES } from 'config'
 import { initShared } from 'shared'
 import { ReportFatalError } from 'shared/loading/ReportFatalError'
 import { defaultLogger } from 'shared/logger'
-import { initializeEngine } from './dcl'
+import { initializeEngine } from './initializeEngine'
 import { Session } from 'shared/session'
 import { waitingForRenderer } from 'shared/loading/types'
 const queryString = require('query-string')
